requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pycryptodome>=3.19.0
slowapi>=0.1.9
limits>=3.7.0
pydantic>=2.5.0
python-multipart>=0.0.6
httpx>=0.27.0
